+++ frontend/src/components/Main/HomeScreen.tsx
@@ -2,7 +2,7 @@ import { useNavigate } from "react-router-dom";
 import { ScrollArea } from "@mantine/core";
 import { useDispatch, useSelector } from "react-redux";
 import { toast } from "react-toastify";
-import LoginForm from "../LandingPage/LoginForm";
+import LoginScreen from "../LandingPage/LoginScreen";
 import Splash from "../LandingPage/Splash";
 
 const HomeScreen = () => {
@@ -14,7 +14,7 @@ const HomeScreen = () => {
   return (
     <ScrollArea h={1000} w="100vw">
       <Splash />
-      <LoginForm />
+      <LoginScreen />
     </ScrollArea>
   );
 };
